// api/src/ai/tools/schema.ts
export const RawFieldItemInputSchema = z.object({
    field: z.string(),
    type: z.string(),
    name: z.string().optional(),
    children: z.union([z.array(z.record(z.string(), z.any())), z.null()]).optional(),
    collection: z.string().optional(),
    schema: z.union([z.record(z.string(), z.any()), z.null()]).optional(),
    meta: z.union([z.record(z.string(), z.any()), z.null()]).optional(),
});

export const RawFieldItemValidateSchema = RawFieldItemInputSchema;

export const FieldItemInputSchema = z.object({
    field: z.string(),
    type: z.string().nullable(),
    name: z.string().optional(),
    collection: z.string().optional(),
    schema: z.union([z.record(z.string(), z.any()), z.null()]).optional(),
    meta: z.union([z.record(z.string(), z.any()), z.null()]).optional(),
});

export const FieldItemValidateSchema = FieldItemInputSchema;